+++ database/migrations/2023_03_28_181518_create_companies_table.php
@@ -17,7 +17,7 @@ class CreateCompaniesTable extends Migration
             $table->id();
             $table->string('name');
             $table->unsignedBigInteger('document_number')->unique();
-            $table->integer('phone')->nullable();
+            $table->bigInteger('phone')->nullable();
             $table->string('email')->unique();
             $table->string('password');
             $table->string('website')->nullable();
